Review

+++ app/controllers/administration/clients/client_admins_controller.rb
@@ -7,11 +7,12 @@ module Administration
       prepend_before_action :set_resource_class
       before_action :ensure_client
       before_action :set_resource, only: %i[show edit update destroy toggle_status sidebar spoof reset_password]
-      before_action :skip_authorization, only: [:sidebar]
+      before_action :skip_authorization, only: %i[sidebar index create edit update destroy]
       append_before_action :init_breadcrumbs, except: %i[new create assign_multiple]
-      append_before_action :pundit_authorize, except: [:sidebar]
+      append_before_action :pundit_authorize, except: %i[sidebar index create edit update destroy]
 
       def index
+        authorize resource_class, :can_manage_client_admins?
         @filter_term = params.dig(:q, :filterable_fields)
         @_filter_form = policy_scope(resource_class).
                         includes(user: %i[clients memberships]).
@@ -44,6 +45,7 @@ module Administration
       end
 
       def create
+        authorize resource_class, :can_manage_client_admins?
         Memberships::CreateAdminCommand.
           call(resource_class.new(create_resource_params), client, current_user, Membership::CLIENT_ADMIN_ROLE) do
           on(:invalid) { render :new, locals: { is_new: true } }
@@ -65,12 +67,14 @@ module Administration
 
       # GET /administration/resources/1/edit
       def edit
+        authorize resource_class, :can_manage_client_admins?
         add_breadcrumb t('administration.breadcrumbs.client_admins'), action: :index
         add_breadcrumb resource.decorate.display_name, action: :edit, id: resource.id
       end
 
       # PATCH/PUT /administration/resources/1
       def update
+        authorize resource_class, :can_manage_client_admins?
         resource.user.modified_by_id = current_user.id
         respond_to do |format|
           if resource.update(update_resource_params)
@@ -86,6 +90,7 @@ module Administration
       end
 
       def destroy
+        authorize resource_class, :can_manage_client_admins?
         if resource.user.memberships.count == 1
           resource.user.destroy
         else
@@ -112,7 +117,8 @@ module Administration
       # Change resources's status to active/disabled
       #
       def toggle_status
-        resource.toggle!(:disabled)
+        authorize resource_class, :can_manage_client_admins?
+        resource_class.update(@_resource.id, disabled: !@_resource.membership_disabled)
         # Reload with join_user
         @_resource = policy_scope(resource_class).join_user.find(params[:id])
         respond_to do |format|
